move HtmlHelperExtension.cs to Slice.Web.Common
