+++ Slice.RecipeWeb/Slice.RecipeWeb.csproj
@@ -128,7 +128,9 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
-    <Compile Include="Common\HtmlHelperExtension.cs" />
+    <Compile Include="Common\HtmlHelperExtension.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="Controllers\BaseController.cs" />
     <Compile Include="Controllers\ContactController.cs" />
     <Compile Include="Controllers\ContentBaseController.cs" />
