+++ Slice.Web/Slice.Web.csproj
@@ -398,6 +398,7 @@
     <Content Include="Views\Recipe\Category.cshtml" />
     <Content Include="Views\Shared\_ReferenceKeywords.cshtml" />
     <Content Include="Views\Navigate\MainMenu.cshtml" />
+    <Content Include="Views\Shared\_SocialShareBar.cshtml" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
